refactor: `UserDictWord` を移動
