+++ voicevox_engine/user_dict/user_dict.py
@@ -10,11 +10,11 @@ from uuid import UUID, uuid4
 import numpy as np
 import pyopenjtalk
 
-from ..model import UserDictWord
 from ..utility.path_utility import get_save_dir, resource_root
 from .part_of_speech_data import (
     MAX_PRIORITY,
     MIN_PRIORITY,
+    UserDictWord,
     WordTypes,
     part_of_speech_data,
 )
